requirements.txt
fastapi == 0.70.0

uvicorn == 0.15.0

motor == 2.5.1

pymongo[tls,srv] == 3.12.1

python-decouple == 3.3

python-multipart == 0.0.5

bcrypt == 3.2.0

pyjwt == 2.3.0
